finops-cost-data-analyst/requirements.txt
google-adk>=1.14
google-cloud-aiplatform[adk,agent-engines]>=1.93.0
google-cloud-bigquery>=3.12.0
toolbox-core>=0.3.0
python-dotenv>=1.0.1
pydantic>=2.11.3
pytest>=8.3.5
pytest-asyncio>=0.26.0
